+++ init.rb
@@ -25,7 +25,7 @@ Redmine::Plugin.register :redmine_oauth do
   name 'Redmine OAuth plugin'
   author 'Karel Pičman'
   description 'Redmine OAuth plugin'
-  version '2.1.5'
+  version '2.2.0'
   url 'https://github.com/kontron/redmine_oauth'
   author_url 'https://github.com/kontron/redmine_oauth/graphs/contributors'
 
@@ -37,6 +37,13 @@ Redmine::Plugin.register :redmine_oauth do
     client_id: '',
     client_secret: '',
     tenant_id: '',
+    custom_name: '',
+    custom_auth_endpoint: '',
+    custom_token_endpoint: '',
+    custom_profile_endpoint: '',
+    custom_scope: 'openid profile email',
+    custom_uid_field: 'preferred_username',
+    custom_email_field: 'email',
     button_color: '#ffbe6f',
     button_icon: 'fas fa-address-card'
   }, partial: 'settings/oauth_settings'